+++ lib/src/commands/commands.dart
@@ -3,5 +3,7 @@ export 'doctor_command.dart';
 export 'flutter_switch_command.dart';
 export 'install_commands/install_command.dart';
 export 'run_command.dart';
+export 'status_command.dart';
+export 'test_command.dart';
 export 'uninstall_command.dart';
 export 'update_command.dart';